# ai-docs/engine/sre-api.openapi.yml
openapi: 3.1.0
info:
  title: Saigon Rider Reward Engine (SRE) API
  version: "1.0.0"
  description: |
    Saigon Rider Reward Engine 의 REST API 명세.
    - 외부(모바일 앱 백엔드 게이트웨이)용 v1 API와 관리자(Admin) API를 함께 정의합니다.
    - 모든 시각은 `RFC 3339` (UTC) 형식, `timestamptz` 와 1:1 매핑.
    - 모든 변동성 API(POST/PUT/DELETE)는 `Idempotency-Key` 헤더로 멱등성을 강제합니다.
    - 에러 응답은 RFC 7807 (`application/problem+json`) 을 따릅니다.
servers:
  - url: https://sre.saigonrider.example.com/v1
    description: Production
  - url: https://sre-stg.saigonrider.example.com/v1
    description: Staging
security:
  - GatewayToken: []
...
paths:
...
  /catalog:
    get:
      tags: [rewards]
      summary: 보상 카탈로그 목록
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum: [BADGE, TELCO, COSMETIC, GIFTCARD]
        - in: query
          name: partner_code
          schema: { type: string }
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/CatalogItem" }
  /catalog/{catalog_id}:
    get:
      tags: [rewards]
      summary: 카탈로그 단건 상세
      parameters:
        - in: path
          name: catalog_id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogItem" }
...
  /admin/catalog:
    post:
      tags: [admin]
      summary: 카탈로그 항목 추가
      security: [ { AdminToken: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CatalogItem" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogItem" }
  /admin/catalog/{catalog_id}:
    put:
      tags: [admin]
      summary: 카탈로그 항목 갱신
      security: [ { AdminToken: [] } ]
      parameters:
        - in: path
          name: catalog_id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CatalogItem" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CatalogItem" }
components:
  securitySchemes:
    GatewayToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        앱 백엔드 게이트웨이가 SRE에 발급받은 서비스 토큰.
        SRE는 게이트웨이의 토큰만 신뢰하며 모바일 앱이 직접 호출하는 것은 금지됩니다.
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 관리자 콘솔 토큰. RBAC 클레임 포함 필요.
  parameters:
...
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
      description: 다음 페이지 커서 (이전 응답의 next_cursor 그대로 전달)
    Limit:
      in: query
      name: limit
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
...
  schemas:
...
    Page:
      type: object
      properties:
        next_cursor: { type: string, nullable: true }
        has_more:    { type: boolean }
...
    CatalogItem:
      type: object
      required: [partner_id, item_code, item_name, category_code, required_rp]
      properties:
        catalog_id:     { type: integer, format: int64, readOnly: true }
        partner_id:     { type: integer, format: int64 }
        partner_code:   { type: string, readOnly: true }
        item_code:      { type: string }
        item_name:      { type: string }
        category_code:
          type: string
          enum: [BADGE, TELCO, COSMETIC, GIFTCARD]
        required_rp:    { type: integer, minimum: 1 }
        face_value_vnd: { type: integer, nullable: true }
        monthly_quota:  { type: integer, nullable: true }
        monthly_issued: { type: integer, readOnly: true }
        is_active:      { type: boolean, default: true }
        visible_from:   { type: string, format: date-time, nullable: true }
        visible_until:  { type: string, format: date-time, nullable: true }
...